Unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/jexl/branches/2.0@810884 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/jexl/JexlArithmetic.java
@@ -158,7 +158,7 @@ public class JexlArithmetic {
             // coerce to int if possible
             long l = bigi.longValue();
             // coerce to int when possible (int being so often used in method parms)
-            if (l <= ((long) Integer.MAX_VALUE) && l >= ((long) Integer.MIN_VALUE)) {
+            if (l <= Integer.MAX_VALUE && l >= Integer.MIN_VALUE) {
                 return new Integer((int) l);
             }
             return new Long(l);
